Ensure order of agent names is preserved
This fixes an issue where the order of agent names was not preserved due to using a set.
This names are being used when preparing the tool node, and are present on the final prompt to the LLM.
On every run, the order was different causing issues when using `LANGSMITH_TEST_CACHE=path/to/cassets`.